+++ dotnet/Core/Repository/Domain/Custom/Database/Special/Person.cs
@@ -58,6 +58,8 @@ namespace Allors.Repository
 
         public int UserAccessFailedCount { get; set; }
 
+        public bool IsDisabled { get; set; }
+
         public Login[] Logins { get; set; }
 
         #endregion